Basic Btree page delete implementation
Get cell_idx from stack and drops the page

Currently the cell_idx we are getting is wrong (`1` instead of `0` for a 1 page Btree). I suspect the problem is with
`seek` which is advancing the stack's cell_indices[0] before `delete` is called. Check test/src/lib.rs::test_simple_delete() unit test.

If I hardcore cell_idx as `0` in delete, delete works.